Move `CopyrightHolders()` and `LicenseInfo()` into `libbitcoin_common`
After https://github.com/bitcoin/bitcoin/pull/26645, this change is
required to be able to `#include clientversion.h` in the
`libbitcoinconsensus` code without dependency on `util/translation.h`.
