+++ src/bitcoin-cli.cpp
@@ -9,6 +9,7 @@
 
 #include <chainparamsbase.h>
 #include <clientversion.h>
+#include <common/license.h>
 #include <common/url.h>
 #include <compat/compat.h>
 #include <compat/stdin.h>
