+++ src/bitcoin-util.cpp
@@ -11,6 +11,7 @@
 #include <chainparams.h>
 #include <chainparamsbase.h>
 #include <clientversion.h>
+#include <common/license.h>
 #include <compat/compat.h>
 #include <core_io.h>
 #include <streams.h>
