+++ src/clientversion.h
@@ -41,11 +41,6 @@ extern const std::string CLIENT_NAME;
 std::string FormatFullVersion();
 std::string FormatSubVersion(const std::string& name, int nClientVersion, const std::vector<std::string>& comments);
 
-std::string CopyrightHolders(const std::string& strPrefix);
-
-/** Returns licensing information (for -version) */
-std::string LicenseInfo();
-
 #endif // WINDRES_PREPROC
 
 #endif // BITCOIN_CLIENTVERSION_H
